Further speed up instruction fetch
Dataflow through the PC is an ILP bottleneck.  Reduce the critical code path
by memoizing the likely next PC.

The first time an instruction is cached, the assumption is that the likely
next PC is on the sequential path.  Whenever the instruction is executed and
the next PC was mispredicted, correct it.  This does the right thing for
jumps and heavily biased branches.  The misprediction penalty is low enough
that doing something smarter for less-biased branches is unprofitable.